+++ internal/utils/context.go
@@ -3,6 +3,7 @@ package utils
 import "time"
 
 type Context struct {
+	DryRun         bool
 	Duration       time.Duration
 	DurationOffset time.Duration
 	Verbose        bool